Properly handle invalid sequences
If the sequence is unrecognized, exit 0 instead of trying to get it from
the array, trying to type the empty string, and then typing a space
(which is what it did before).

+++ morse_code_open.sh
@@ -95,6 +95,9 @@ fi
 
 sleep "$letter_pause"
 sequence=$(cat /tmp/morse_code_letter)
+if [[ ! -v "morse_letters[$sequence]" ]] ; then
+    exit 0
+fi
 xdotool type "${morse_letters[$sequence]}"
 rm /tmp/morse_code_letter
